requirements.txt
requests==2.21.0
green==2.13.0
flask==1.0.2
SQLAlchemy>=1.3.0
green==2.13.0
psycopg2-binary==2.7.5
gunicorn==19.5.0
gevent==1.2.1
python-dateutil==2.8.0
uWSGI==2.0.17.1
werkzeug==0.16.1
